+++ src/promptflow/tests/test_configs/flows/chat_group/cloud_batch_runs/chat_group_copilot/flow.dag.yaml
@@ -1,0 +1,34 @@
+$schema: https://azuremlschemas.azureedge.net/promptflow/latest/Flow.schema.json
+inputs:
+  question:
+    type: string
+    default: What's human?
+  model:
+    type: string
+    default: gpt-3.5-turbo
+  conversation_history:
+    type: list
+outputs:
+  output:
+    type: string
+    reference: ${llm.output}
+nodes:
+- name: prompt
+  type: prompt
+  inputs:
+    question: ${inputs.question}
+    conversation_history: ${inputs.conversation_history}
+  source:
+    type: code
+    path: prompt.jinja2
+- name: llm
+  type: llm
+  inputs:
+    prompt: ${prompt.output}
+    deployment_name: gpt-35-turbo
+    model: ${inputs.model}
+  source:
+    type: code
+    path: prompt.jinja2
+  connection: azure_open_ai_connection
+  api: chat
