+++ src/promptflow/tests/test_configs/flows/chat_group/cloud_batch_runs/chat_group_simulation/flow.dag.yaml
@@ -1,0 +1,22 @@
+$schema: https://azuremlschemas.azureedge.net/promptflow/latest/Flow.schema.json
+inputs:
+  topic:
+    type: string
+  ground_truth:
+    type: string
+  conversation_history:
+    type: list
+outputs:
+  output:
+    type: string
+    reference: ${simulator.output}
+nodes:
+- name: simulator
+  type: python
+  source:
+    type: code
+    path: simulator.py
+  inputs:
+    question: ${inputs.topic}
+    ground_truth: ${inputs.ground_truth}
+    conversation_history: ${inputs.conversation_history}
