+++ src/promptflow/tests/test_configs/flows/chat_group_simulation/simulator.py
@@ -1,30 +1,28 @@
-import uuid
 from typing import List
 
 from promptflow import tool
 
 
-def get_question_from_conversation_history(conversation_history: List, question: str) -> str:
+def get_answer_from_conversation_history(conversation_history: List) -> str:
     """
     This function gets the answer from the conversation history.
 
     :param conversation_history: the conversation history.
     """
     if len(conversation_history) == 0:
-        return question
-    generated_question = f"{question}_{str(uuid.uuid4())}"
-    return generated_question
+        return "NA"
+    assistant_answers = [item[1] for item in conversation_history if item[0] == "assistant"]
+    return assistant_answers[-1]["output"]
 
 
 @tool
 def simulate(question: str, ground_truth: str, conversation_history: List) -> str:
     print(f"question: {question}")
     print(f"chat_history: {conversation_history}")
-    generated_question = get_question_from_conversation_history(conversation_history, question)
-    print(f"generated_question: {generated_question}")
-    if generated_question != ground_truth:
-        return generated_question
+    answer = get_answer_from_conversation_history(conversation_history)
+    print(f"answer: {answer}")
+    if answer != ground_truth:
+        return "I don't like this answer, give me another one."
     else:
         return "[STOP]"
-
